[all]: Added MOVW,MOVT,BX, adapted parser
Motivation: Test the compilation of Armv7 programs, For which we need
some more instructions and to be able to parse instruction tokens in an
ARM ELF Grammar.

This patch does two things:
- Adds support for MOVW,MOVT,and BX instructions
- Updates ARMLexer so that instruction tokens are defined in
  check_names. This matches the design of AArch64Parser and is needed so
we can export instruction tokens and reuse them in other Grammars.

Tests are provided and pass in herd, however currently both tests do not
work in litmus:

- A013 produces very strange behaviour when run on a Raspberry Pi 2 - is
  this a hardware bug?
- A015 I cannot currently get this test to work as litmus does not
  generate the required ins_t type for this test. I do not know why.