[lib,herd]: Test C06 passes
Fix macro printing of __int128_t and __uint128_t types.
There is not a macro in inttypes for int128 types, so we print directly
as `llx`.

This means when C06 is run on `litmus` you get a warning rather than a
failure.

This is because int128 types are implementation defined as to whether
they are supported - not all architectures support them. e.g. ARM does
not