[all] Add LDM Instruction support
Motivation: Wish to test the compilation of Armv7 programs

This patch adds support for the 'Load multiple' instruction.

This instruction has the form:
```
LDM Ra, { Rx, Ry, Rz }
```
Where Ra contains the address to be loaded from and the number of
registers within the {} determines how many consecutive registers to
load. I implement the 2-register and 3-register variants here.

Two tests (A007 and A008) are provided and pass in both herd and litmus (tested using a Raspberry Pi 2).

I could not test A009 on litmus using the Rasberry Pi 2 as it does not
support the LDM instruction - I do not have newer hardware.