[litmus] Do not require the same argument multipe times
Consider the following litmus test:

AArch64 CMP-NOP
{
 0:X0=NOP; 0:X1=NOP;
}
 P0          ;
 CMP W0,W1   ;
 CSET W2,EQ  ;
exists(0:X2=0)

Prior to this patch, litmus7 would fail to generate valid c sources
for this test due to the fact that we're passing the NOP as a function
argument to code0 twice.

This patch fixes this issue by avoiding the need to pass the same
constant twice.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>