[herd,asl] Fix memory read
In spite of its type (and of an explicit type constraint!) `PhysMemRead`
from implementation.asl was always returning a bitvector of size 64,
even when a shorter memory read was required. The root cause
is the systematic conversion to  bitvector of size 64
by the primitive `read_memory_gen`.

Fix consists in adding a slicing operation. The type of `read_memory_gen`
is also corrected.